// src/types/models/UIStateCode.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bodoge API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const UIStateCode = {
    Start: 'START',
    Initializing: 'INITIALIZING',
    InYourTurn: 'IN_YOUR_TURN',
    InOthersTurn: 'IN_OTHERS_TURN',
    PlayOneCard: 'PLAY_ONE_CARD',
    PlayTwoCard: 'PLAY_TWO_CARD',
    PlayCardToOther: 'PLAY_CARD_TO_OTHER',
    UndoEffect1: 'UNDO_EFFECT_1',
    UndoEffect2: 'UNDO_EFFECT_2',
    PeeqEffect1: 'PEEQ_EFFECT_1',
    PeeqEffect2: 'PEEQ_EFFECT_2',
    YouWon: 'YOU_WON',
    YouLost: 'YOU_LOST',
    Error: 'ERROR',
    Viewer: 'VIEWER',
    End: 'END'
} as const;
export type UIStateCode = typeof UIStateCode[keyof typeof UIStateCode];


export function instanceOfUIStateCode(value: any): boolean {
    for (const key in UIStateCode) {
        if (Object.prototype.hasOwnProperty.call(UIStateCode, key)) {
            if (UIStateCode[key as keyof typeof UIStateCode] === value) {
                return true;
            }
        }
    }
    return false;
}

export function UIStateCodeFromJSON(json: any): UIStateCode {
    return UIStateCodeFromJSONTyped(json, false);
}

export function UIStateCodeFromJSONTyped(json: any, ignoreDiscriminator: boolean): UIStateCode {
    return json as UIStateCode;
}

export function UIStateCodeToJSON(value?: UIStateCode | null): any {
    return value as any;
}

export function UIStateCodeToJSONTyped(value: any, ignoreDiscriminator: boolean): UIStateCode {
    return value as UIStateCode;
}